Improve typing

+++ core/executor/src/wasm_executor.rs
@@ -64,7 +64,7 @@ impl<'e, E: Externalities<Blake2Hasher>> FunctionExecutor<'e, E> {
 
 		Ok(FunctionExecutor {
 			sandbox_store: sandbox::Store::new(),
-			heap: allocator::FreeingBumpHeapAllocator::new(used_size, heap_size, m.clone()),
+			heap: allocator::FreeingBumpHeapAllocator::new(used_size as u32, heap_size as u32, m.clone()),
 			memory: m,
 			table: t,
 			ext: e,
